+++ examples/package/docs/02-Foo/Button.tsx
@@ -1,5 +1,0 @@
-import * as React from 'react';
-
-export default () => {
-  return <button> button test</button>;
-};