+++ examples/package/src/test.ts
@@ -1,1 +1,0 @@
-export const multiply = (a) => (b) => a + b;