+++ examples/react-component/src/index.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+
+export default function Component() {
+  return (
+    <h1>Hello World</h1>
+  );
+}