+++ examples/react-multi-components/src/Avatar/index.tsx
@@ -1,0 +1,15 @@
+import * as React from 'react';
+import defaultAvatar from './default.png';
+import './index.css';
+
+interface AvatarProps {
+
+}
+
+const Avatar: React.FunctionComponent<React.PropsWithChildren<AvatarProps>> = (props) => {
+  return (
+    <img src={defaultAvatar} alt="avatar" className="pkg-avatar" />
+  );
+};
+
+export default Avatar;