+++ packages/pkg/src/build.ts
@@ -1,7 +1,7 @@
 import { mergeConfigOptions } from './helpers/mergeConfigOptions.js';
 import { buildAll } from './buildAll.js';
 
-import type { PkgContext } from './types.js';
+import type { PkgContext, PkgTaskConfig } from './types.js';
 
 export default async (context: PkgContext) => {
   const { getConfig, applyHook, commandArgs } = context;
@@ -10,8 +10,7 @@ export default async (context: PkgContext) => {
   await applyHook('before.build.load', { args: commandArgs, config: configs });
 
   if (!configs.length) {
-    const err = new Error('Could not Find any pending tasks when excuting \'build\' command.');
-    throw err;
+    throw new Error('Could not Find any pending tasks when executing \'build\' command.');
   }
 
   await applyHook('before.build.run', {
@@ -19,13 +18,12 @@ export default async (context: PkgContext) => {
     config: configs,
   });
 
-  // @ts-ignore fixme
-  const normalizedConfigs = configs.map((config) => mergeConfigOptions(config, context));
+  const normalizedConfigs = configs.map((config) => mergeConfigOptions(config as PkgTaskConfig, context));
 
   try {
-    await buildAll(normalizedConfigs, context);
+    const outputResults = await buildAll(normalizedConfigs, context);
 
-    await applyHook('after.build.compile');
+    await applyHook('after.build.compile', outputResults);
   } catch (err) {
     await applyHook('error', {
       errCode: 'COMPILE_ERROR',