+++ packages/pkg/src/helpers/mergeConfigOptions.ts
@@ -1,50 +1,62 @@
 import deepmerge from 'deepmerge';
-import { getEntryDir, getEntryFile, getOutputDir } from './getTaskIO.js';
-import { getBundleSwcConfig, getTransformSwcConfig } from './getSwcConfig.js';
+import { formatEntry, getDefaultEntryDir, getOutputDir } from './getTaskIO.js';
+import { getDefaultBundleSwcConfig, getDefaultTransformSwcConfig } from './defaultSwcConfig.js';
 import { normalizeRollupConfig } from './normalizeRollupConfig.js';
+import { stringifyObject } from '../utils.js';
 
-import type { PkgContext, TaskLoaderConfig, PkgTaskConfig } from '../types.js';
+import type { PkgContext, PkgTaskConfig, TaskLoaderConfig } from '../types.js';
 
 export const mergeConfigOptions = (
   cfg: PkgTaskConfig,
   ctx: PkgContext,
 ): TaskLoaderConfig => {
-  const { rootDir, userConfig } = ctx;
+  const { rootDir, command } = ctx;
   const { config: taskConfig, name: taskName } = cfg;
-  const normalizedConfig = { ...taskConfig };
-  const { type, entry, outputDir, swcCompileOptions = {} } = normalizedConfig;
+  const normalizedConfig = { ...taskConfig, name: taskName };
+  const { type, entry, outputDir, swcCompileOptions = {}, define } = normalizedConfig;
   const isBundleTask = type === 'bundle';
+  const isTransformTask = type === 'transform';
+
+  if (isBundleTask) {
+    normalizedConfig.mode = taskConfig.mode;
+  } else if (isTransformTask) {
+    normalizedConfig.mode = command === 'build' ? 'production' : 'development';
+  }
 
   // Configure task entry
-  normalizedConfig.entry = entry || (
-    isBundleTask ? getEntryFile(rootDir) : getEntryDir(rootDir)
-  );
+  if (isBundleTask) {
+    normalizedConfig.entry = formatEntry(entry);
+  } else if (isTransformTask) {
+    normalizedConfig.entry = getDefaultEntryDir(rootDir);
+  } else {
+    throw new Error('Invalid task type.');
+  }
 
   // Configure task outputDir
   normalizedConfig.outputDir = outputDir || getOutputDir(rootDir, taskName);
 
+  // Configure define
+  normalizedConfig.define = stringifyObject(define || {});
+
+  normalizedConfig.sourcemap = normalizedConfig.sourcemap ?? command === 'start';
+
   // Configure swcOptions
-  const swcOptionOverride = deepmerge(
+  normalizedConfig.swcCompileOptions = deepmerge(
     isBundleTask
-      ? getBundleSwcConfig(userConfig, taskName)
-      : getTransformSwcConfig(userConfig, taskName),
+      ? getDefaultBundleSwcConfig(taskName)
+      : getDefaultTransformSwcConfig(normalizedConfig, taskName),
     swcCompileOptions,
   );
 
-  normalizedConfig.swcCompileOptions = swcOptionOverride;
-
   // Configure rollup plugins & options
-  const [resolvedPlugins, resolvedRollupOption] = normalizeRollupConfig(
+  const [resolvedPlugins, resolvedRollupOptions] = normalizeRollupConfig(
     normalizedConfig,
     ctx,
     taskName,
   );
 
   normalizedConfig.rollupPlugins = resolvedPlugins;
-  normalizedConfig.rollupOptions = resolvedRollupOption;
+  normalizedConfig.rollupOptions = resolvedRollupOptions;
 
-  return {
-    ...normalizedConfig,
-    name: taskName,
-  };
+  return normalizedConfig;
 };