+++ packages/pkg/src/helpers/reportSize.ts
@@ -2,13 +2,13 @@ import consola from 'consola';
 import picocolors from 'picocolors';
 import { gzipSizeSync } from 'gzip-size';
 
-const UNIT = ['B', 'KB', 'MB', 'GB', 'TB'];
+const UNIT = ['B', 'KiB', 'MiB', 'GiB', 'TiB'];
 
 const prettifySize = (bytes: number): string => {
   if (bytes === 0) return '0 B';
 
   const exp = Math.floor(Math.log2(bytes) / 10);
-  return `${(bytes / Math.pow(1024, exp)).toFixed(2)} ${UNIT[exp]}`;
+  return `${(`${(bytes / Math.pow(1024, exp)).toFixed(2)}`).replace(/\.00/, '')} ${UNIT[exp]}`;
 };
 
 const findMaxLength = (names: string[]) => {
@@ -22,11 +22,11 @@ export const reportSize = (
 ) => {
   const names = Object.keys(files);
   const maxLen = findMaxLength(names);
-  const padLength = maxLen > 30 ? (maxLen + 2) : 30;
+  const padLength = maxLen > 35 ? (maxLen + 2) : 35;
 
   names.forEach((name) => {
     const rawSize = prettifySize(files[name].length);
     const gzipSize = prettifySize(gzipSizeSync(files[name]));
-    consola.info(`${name.padStart(padLength, ' ')}: ${picocolors.green('raw')} ${rawSize} ${picocolors.cyan('gzip')} ${gzipSize}`);
+    consola.info(`${name.padStart(padLength, ' ')}: ${rawSize}, ${picocolors.cyan('Gzipped')} ${gzipSize}`);
   });
 };