+++ packages/pkg/src/types.ts
@@ -1,48 +1,153 @@
-import type { RollupOptions, Plugin, SourceMapInput } from 'rollup';
+import type { RollupOptions, Plugin, SourceMapInput, ModuleJSON, RollupOutput } from 'rollup';
 import type { Context, IPluginAPI, IPlugin, ITaskConfig } from 'build-scripts';
 import type { Config } from '@swc/core';
+import type stylesPlugin from 'rollup-plugin-styles';
+
+type StylesRollupPluginOptions = Parameters<typeof stylesPlugin>[0];
 
 export type PlainObject = Record<string, string | boolean | number | object>;
+export type ReverseMap<T> = T[keyof T];
 
 export type RollupPluginFn<T = {}> = (args?: T) => Plugin;
 
-export interface TaskLoaderConfig extends TaskConfig {
+interface TransformOptions {
+  /**
+   * Which type of contents would be generated
+   * "cjs" - Commonjs with ES5 syntax (targeting Node version under 12);
+   * "esm" - ES Module with ES5 syntax (legacy outputs);
+   * "es2017" - ES Module with ES2017 (targeting modern browsers and Node version upon 12)
+   * @default ['esm','es2017']
+   */
+  formats?: Array<'cjs' | 'esm' | 'es2017'>;
+  /**
+   * Exclude all files matching any of those conditions.
+   * - `string` to match any paths by `minimatch` glob patterns
+   * - An `array` to match at least one of the conditions
+   * @see https://github.com/isaacs/minimatch
+   */
+  excludes?: string | string[];
+}
+
+interface BundleOptions {
+  /**
+   * Export name
+   * @default package.name
+   */
+  name?: string;
+  /**
+   * As the name of the generated file.
+   * @default index
+   */
+  filename?: string;
+  /**
+   * Generate uncompressed bundle for development debug.
+   */
+  development?: boolean;
+  /**
+   * Which type of contents would be generated
+   * "umd"
+   * "esm"
+   * "es2017"
+   * @default ['esm','es2017']
+   */
+  formats?: Array<'umd' | 'esm' | 'cjs' | 'es2017'>;
+  /**
+   * Specify external dependencies.
+   * "boolean" - whether to bundle all dependencies or not;
+   * "object" - specific external dependencies.
+   * @default false All of dependencies will be bundled by default.
+   */
+  externals?: boolean | Record<string, string>;
+
+  /**
+   * Minify JS and CSS bundle.
+   */
+  minify?: boolean;
+}
+
+export type TaskLoaderConfig = BundleTaskLoaderConfig | TransformTaskLoaderConfig;
+
+export interface BundleTaskLoaderConfig extends BundleTaskConfig {
+  type: 'bundle';
   name: TaskName;
 }
 
-export interface TaskConfig {
+export interface TransformTaskLoaderConfig extends TransformTaskConfig {
+  type: 'transform';
+  name: TaskName;
+}
+
+interface CommonTaskConfig {
   /**
-   * There are two ways to handle source files
-   * - 'transform' to transform files one by one to support tree shaking
-   * - 'bundle' to bundle files into single file
+   * Build mode.
    */
-  type: 'bundle' | 'transform';
-  /**
-  * Entry for a specific task
-  * @default `./src` for transforming task, and `./src/index[j|t]s` for bundling task
-  */
-  entry?: string;
+  mode?: NodeEnvMode;
   /**
   * Output directory
   */
   outputDir?: string;
   /**
-  * Extra rollup plugins
-  */
-  rollupPlugins?: Plugin[];
-
+   * Define global constant replacements
+   */
+  define?: Record<string, string>;
+  /**
+    * - true to generate a sourcemap for the code and include it in the result object.
+    * - "inline" to generate a sourcemap and append it as a data URL to the end of the code,
+    * but not include it in the result object.
+    */
+  sourcemap?: boolean | 'inline';
+  /**
+   *  Alias to file system paths
+   */
+  alias?: Record<string, string>;
   /**
   * Extra rollup options
   * @see https://rollupjs.org/guide/en/
   */
   rollupOptions?: RollupOptions;
-
+  /**
+  * Extra rollup plugins
+  */
+  rollupPlugins?: Plugin[];
   /**
   * Extra swc compile options
   * @see https://swc.rs/docs/configuration/compilationv
   */
   swcCompileOptions?: Config;
+  /**
+   * Extra babel plugins
+   */
+  babelPlugins?: babel.PluginItem[];
 }
+
+export interface BundleTaskConfig extends CommonTaskConfig, BundleOptions {
+  type: 'bundle';
+  /**
+  * Entry for a specific task
+  * @default  `./src/index[j|t]sx` for bundling task
+  */
+  entry?: RollupOptions['input'];
+  /**
+  * Files extensions
+  * @see https://www.npmjs.com/package/@rollup/plugin-node-resolve
+  */
+  extensions?: string[];
+  /**
+   * Config styles options. See https://www.npmjs.com/package/rollup-plugin-styles
+   */
+  stylesOptions?: (options: StylesRollupPluginOptions) => StylesRollupPluginOptions;
+}
+
+export interface TransformTaskConfig extends CommonTaskConfig, TransformOptions {
+  type: 'transform';
+  /**
+  * Entry for a specific task
+  * @default  `./src` for Transform task
+  */
+  entry?: string;
+}
+
+export type TaskConfig = BundleTaskConfig | TransformTaskConfig;
 
 export type PkgTaskConfig = ITaskConfig<TaskConfig, TaskName>;
 
@@ -77,17 +182,28 @@ export interface OutputFile {
   map?: string | SourceMapInput;
 }
 
+export interface OutputResult {
+  taskName: string;
+  outputFiles: OutputFile[];
+  // Only exist in rollup bundle task
+  modules?: ModuleJSON[];
+  outputs?: Array<RollupOutput['output']>;
+}
+
 export interface UserConfig {
   /**
+   * Entry for build.
+   */
+  entry?: string | string[] | Record<string, string>;
+  /**
    * Alias to file system paths
-   * @default empty
+   * @default {}
    */
   alias?: Record<string, string>;
   /**
    * Define global constant replacements
-   * @default empty
    */
-  define?: PlainObject;
+  define?: Record<string, string>;
   /**
   * - true to generate a sourcemap for the code and include it in the result object.
   * - "inline" to generate a sourcemap and append it as a data URL to the end of the code,
@@ -102,61 +218,20 @@ export interface UserConfig {
 
   /**
    * Plugins of build scripts
-   * @default @ice/pkg-plugin-component
+   * @default []
    */
-  plugins?: any;
+  plugins?: Array<string | [string, any?]>;
 
   /**
    * "transform mode" means transform files one by one
    */
-  transform?: {
-    /**
-     * Which type of contents would be generated
-     * "cjs" - Commonjs with ES5 synatx (targeting Node version under 12);
-     * "esm" - ES Module with ES5 synatx (lagacy outputs);
-     * "es2017" - ES Module with ES2017 (targeting modern browsers and Node version upon 12)
-     * @default ['esm','es2017']
-     */
-    formats?: Array<'cjs' | 'esm' | 'es2017'>;
-    /**
-     * Exclude all files matching any of those conditions.
-     * - `string` to match any paths by `minimatch` glob patterns
-     * - An `array` to match at least one of the conditions
-     * @see https://github.com/isaacs/minimatch
-     */
-    excludes?: string | string[];
-  };
+  transform?: TransformOptions;
 
   /**
-   * "bundle mode" means bundle everything up by using of Rollup
+   * "bundle mode" means bundle everything up by using Rollup
    */
-  bundle?: {
-    /**
-     * Export name
-     * @default package.name
-     */
-    name?: string;
-    /**
-     * As the name of the generated file.
-     * @default index
-     */
-    filename?: string;
-
-    development?: boolean;
-    /**
-     * Which type of contents would be generated
-     * "umd"
-     * "esm"
-     * "es2017"
-     * @default ['esm','es2017']
-     */
-    formats?: Array<'umd' | 'esm' | 'es2017'>;
-    /**
-     * Specify external dependencies.
-     * "boolean" - whether to bundle all dependencies or not;
-     * "object" - specific external dependencies.
-     * @default true nono of dependencies will be bundled by default
-     */
-    externals?: boolean | Record<string, string>;
-  };
+  bundle?: BundleOptions;
 }
+
+// Set for `process.env.NODE_ENV`
+export type NodeEnvMode = 'development' | 'production' | string;