+++ packages/plugin-docusaurus/src/plugin.js
@@ -78,8 +78,6 @@ module.exports = function (context) {
             'react/jsx-runtime': path.resolve(siteDir, 'node_modules/react/jsx-runtime'),
             '@mdx-js/react': mdxReactPath,
             'style-unit': styleUnitPath,
-            // FIXME: I am not sure how to resolve the actual output folder
-            [pkgMeta.name]: path.resolve(siteDir, 'esm/index'),
           }, raxAlias),
         },
         plugins: [